Included LM317 voltage regulator for Beagle Bone
This is the power supply for it, removing the need for a power adapter
for it.

This also changes the AC input to needing a dual winding transformer. My
idea of univeral adapters was nice, until AC "wall warts" became illegal
because of energy laws... Plus, I reduced the supply caps, so I need the
full-wave rectifier now.

I also removed the jumpers to choose to feed the Beagle Bone from the
analog power supply. This was an idea that was never going to work, it
was clear early on. Current demand of one LM317 was too high, and it
would make the analog supply noisy.