Separate AGND plane in two, to avoid hum
The one large ground plane seemed to cause the charging current for the
main filter caps to circulate around the entire plane, causing the
signal output ground and circuit ground reference to have different
potentials, causing a 100 Hz max hum of -75dBV on some outputs.

This fix breaks the ground plane in two and connects them together with
a small track, to keep the circuit ground clear of those charging
currents.

I wonder if even a 1 Ohm resistor should be used to connect the power
supply ground plane to the circuit ground plane.

This fix hasn't been tested at this point, but it's likely what's
needed.

In the prototype I fixed the hum by cutting the input connector grounds
loose from the ground plane, and wiring them with a dedicated wire to
the circuit ground reference point.