Add another power supply cap to fix boot problems
The extra cap (C17) is 2200µF.

The problem was that it sometimes refused to boot. I saw the
BeagleBone's blue lights go off again when the boot process was
initiated. The LCD backlight dipped. Scope image revealved dips in the
rail, together with 100Hz rectification ripple. As a hack to my build, I
added 4700µF before the positive regulators.  That prevented the
dips. I'm not sure if this one extra 2200µF is enough though. Probably
it is, and there are space constraints.